2026_March/requirements.txt
google-api-python-client>=2.0.0
google-auth-oauthlib>=0.5.0
google-auth-httplib2>=0.1.0
Pillow>=10.0.0
